Terminando o inicio
