+++ Botao-Do-Bem_Api.sln
@@ -2,7 +2,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pokedex", "Pokedex.csproj", "{4C5465EE-87F3-0393-CCA8-ADC7154948E8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Botaodobem", "Botaodobem\Botaodobem.csproj", "{769C62A6-30EE-B0BB-0DBA-6CA62D7D3BCC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,15 +10,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4C5465EE-87F3-0393-CCA8-ADC7154948E8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4C5465EE-87F3-0393-CCA8-ADC7154948E8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4C5465EE-87F3-0393-CCA8-ADC7154948E8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4C5465EE-87F3-0393-CCA8-ADC7154948E8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{769C62A6-30EE-B0BB-0DBA-6CA62D7D3BCC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{769C62A6-30EE-B0BB-0DBA-6CA62D7D3BCC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{769C62A6-30EE-B0BB-0DBA-6CA62D7D3BCC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{769C62A6-30EE-B0BB-0DBA-6CA62D7D3BCC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B5C4446E-C0CD-48EF-93B5-A95F1D231C42}
+		SolutionGuid = {663135B1-2EC6-445E-9EAC-3E3A62CFDB96}
 	EndGlobalSection
 EndGlobal
