--- BotaoDoBemApi/BotaoDoBemApi.cs
@@ -1,10 +1,0 @@
-namespace BotaoDoBemApi
-{
-    public class Emocao
-    {
-        public int Id {get; set;}
-        public string sentimento {get; set;}
-        public string frase {get; set; }
-
-    }
-}
