--- BotaoDoBemApi/BotaoDoBemApi.http
@@ -1,6 +1,0 @@
-@BotaoDoBemApi_HostAddress = http://localhost:5255
-
-GET {{BotaoDoBemApi_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
