--- BotaoDoBemApi/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>Pokedex</title>
-</head>
-<body>
-    <h1>Pokedex da alegria</h1>
-
-    <ul id="lista-pokemons"></ul>
-    <script src="./script.js"></script>
-</body>
-</html>
