--- BotaoDoBemApi/style.css
@@ -1,3 +1,0 @@
-.xalala {
-    background-color: crimson;
-}
