--- BotaoDoBemApi/xablau.sql
@@ -1,7 +1,0 @@
-INSERT INTO pokemons (Nome, Peso)
-VALUES 
-  ('Pikachu', 6.0),
-  ('Gengar', 40.0),
-  ('Lapras"', 220),
-  ('Snorlax', 460),
-  ('Squirtle', 9.13);
