+++ Botaodobem/AppDbContext.cs
@@ -1,0 +1,10 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace Botaodobem;
+
+public class AppDbContext : DbContext
+{
+    public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
+
+    public DbSet<Emocao> Emocoes { get; set; }
+}
