+++ Botaodobem/Emocao.cs
@@ -1,0 +1,8 @@
+namespace Botaodobem;
+
+public class Emocao
+{
+    public int Id { get; set; }
+    public string? Nome { get; set; }
+    public string? Frase { get; set; }
+}
