+++ Botaodobem/Program.cs
@@ -1,0 +1,26 @@
+using Microsoft.EntityFrameworkCore;
+using Botaodobem;
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddDbContext<AppDbContext>(options =>
+    options.UseSqlite("Data Source=app.db"));
+
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.MapGet("/weatherforecast", () => "Tudo certo!")
+    .WithName("GetWeatherForecast")
+    .WithOpenApi();
+
+app.Run();
